initial jest config

+++ scripts/test.bash
@@ -11,6 +11,11 @@ if [ $? != 0 ]; then
 	exit 1
 fi
 
+./node_modules/.bin/jest
+if [ $? != 0 ]; then
+    exit 1
+fi
+
 rm *.svg
 rm *.pb
